+++ cdap-app-templates/cdap-etl/cdap-etl-api/src/main/java/co/cask/cdap/templates/etl/api/StageContext.java
@@ -17,13 +17,12 @@
 package co.cask.cdap.templates.etl.api;
 
 import co.cask.cdap.api.metrics.Metrics;
-import co.cask.cdap.api.templates.AdapterContext;
 import co.cask.cdap.api.templates.plugins.PluginProperties;
 
 /**
  * Context passed to ETL stages.
  */
-public interface StageContext extends AdapterContext {
+public interface StageContext {
 
   /**
    * Gets the {@link PluginProperties} associated with the stage.
