+++ cdap-app-templates/cdap-etl/cdap-etl-lib/src/test/java/co/cask/cdap/templates/etl/transforms/MockStageContext.java
@@ -17,14 +17,12 @@
 package co.cask.cdap.templates.etl.transforms;
 
 import co.cask.cdap.api.metrics.Metrics;
-import co.cask.cdap.api.templates.AdapterSpecification;
 import co.cask.cdap.api.templates.plugins.PluginProperties;
 import co.cask.cdap.templates.etl.api.StageContext;
 import co.cask.cdap.templates.etl.common.NoopMetrics;
 import com.google.common.collect.Maps;
 
 import java.util.Map;
-import javax.annotation.Nullable;
 
 /**
  * Mock context for unit tests
@@ -48,26 +46,5 @@ public class MockStageContext implements StageContext {
   @Override
   public Metrics getMetrics() {
     return NoopMetrics.INSTANCE;
-  }
-
-  @Nullable
-  @Override
-  public AdapterSpecification getAdapterSpecification() {
-    return null;
-  }
-
-  @Override
-  public PluginProperties getPluginProperties(String pluginId) {
-    return null;
-  }
-
-  @Override
-  public <T> Class<T> loadPluginClass(String pluginId) {
-    return null;
-  }
-
-  @Override
-  public <T> T newPluginInstance(String pluginId) throws InstantiationException {
-    return null;
   }
 }
